Support alternative SPI backends and allow the default to be changed
This makes it possible to use other SPI controller implementations and
not just the Linux SPI dev implementation. That means that programs and
libraries using Circuits.SPI can be made to work on non-Nerves devices,
interact with simulated SPI devices, and be easier to unit test.

New backends should implement the `Circuits.SPI.Backend`. The SPI dev
implementation does this and is the default. Backends return
`Circuits.SPI.Bus` implementations when opened so SPI calls get
dispatched to the right place without any code changes to user programs.
In fact, this change is mostly invisible to Circuits.SPI users.

To change the default backend, add the following to the `config.exs`:

config :circuits_spi, default_backend: MyOwnSPIBackend

Circuits.SPI currently won't compile the Linux SPI dev backend if you do
this. This means that if you're not running on Nerves or Linux device,
you don't have to worry about Linux bits being compiled.